Add the handler implementation for the delete note endpoint

+++ internal/database/query.sql.go
@@ -138,13 +138,24 @@ func (q *Queries) DeleteAccessTokenByUserId(ctx context.Context, userID uuid.UUI
 	return id, err
 }
 
-const deleteNoteById = `-- name: DeleteNoteById :exec
-DELETE FROM notes WHERE id = $1
+const deleteNoteById = `-- name: DeleteNoteById :one
+DELETE FROM notes WHERE id = $1 RETURNING id, user_id, title, content, background_color, create_time, update_time, delete_time
 `
 
-func (q *Queries) DeleteNoteById(ctx context.Context, id uuid.UUID) error {
-	_, err := q.db.ExecContext(ctx, deleteNoteById, id)
-	return err
+func (q *Queries) DeleteNoteById(ctx context.Context, id uuid.UUID) (Note, error) {
+	row := q.db.QueryRowContext(ctx, deleteNoteById, id)
+	var i Note
+	err := row.Scan(
+		&i.ID,
+		&i.UserID,
+		&i.Title,
+		&i.Content,
+		&i.BackgroundColor,
+		&i.CreateTime,
+		&i.UpdateTime,
+		&i.DeleteTime,
+	)
+	return i, err
 }
 
 const deleteRefreshTokenByUserId = `-- name: DeleteRefreshTokenByUserId :one
